Refactor graphicsmagick packages and consolidate mime-related capabilities.
