+++ src/main/java/com/afrozaar/util/graphicsmagick/meta/MetaDataFormat.java
@@ -1,4 +1,4 @@
-package com.afrozaar.util.graphicsmagick;
+package com.afrozaar.util.graphicsmagick.meta;
 
 import java.util.Optional;
